remove files with PY3

+++ prody/tests/proteins/test_insty.py
@@ -153,8 +153,9 @@ class TestInteractions(unittest.TestCase):
 
     @classmethod
     def tearDownClass(cls):
-        import os
-        for filename in ['test_2k39_all.npy', 'test_2k39_hbs.npy', 'test_2k39_sbs.npy', 
-                         'test_2k39_rib.npy', 'test_2k39_PiStack.npy', 'test_2k39_PiCat.npy',
-                         'test_2k39_hph.npy', 'test_2k39_disu.npy']:
-            os.remove(filename)
+        if PY3K:
+            import os
+            for filename in ['test_2k39_all.npy', 'test_2k39_hbs.npy', 'test_2k39_sbs.npy', 
+                            'test_2k39_rib.npy', 'test_2k39_PiStack.npy', 'test_2k39_PiCat.npy',
+                            'test_2k39_hph.npy', 'test_2k39_disu.npy']:
+                os.remove(filename)
